# Fix performance regression in `previous_version` function